Day 6 and 7: Created Hangman
Completed Days 6 and 7 of the course, which involved the creation of a Hangman application and a series of exercises focusing on functions, loops, and packages. The Hangman application presented a practical, hands-on opportunity to apply these concepts. Additionally, exercises were carried out on the Reeborg's World platform (https://reeborg.ca), further reinforcing the day's lessons.